Import the models from Blender
After creating and texturing the 3D models in Blender and after rigging and animating the character and the ghost models, it’s time to export them from Blender to Unity. The models were exported as FBX files. The house models were created in three LOD versions each, and the LOD0, LOD1 and LOD2 distances were adjusted in Unity.